extended mode
This change makes tweet_mode extended the new default for twarc. The
reasoning behind this is that it turns out not all media entities are
included in compat mode which is highly problematic for data collection.
Whatever additional difficulty there is in parsing this content is minimal
compared to the difficulty presented by missing data.

For users that want it they can set tweet_mode=compat still.

Fixes #168